feat(data): change points/coins data type from Int to Long
Changes the data type for user points and coins from `Int` to `Long` across various data models, requests, and UI components. This is to support larger point values and prevent potential integer overflow issues.

This affects the following areas:
- `UserListResponse`, `UserDetailResponse`, `CreateUserRequest`, and `UpdateUserRequest` in the `manage-users` feature.
- `UserMeResponse` and related UI components (`CtaYourPoint`, `HomeScreen`) in the `home` feature.
- The core `UserData` model and `UserPointFlow` handler.
- The `ExchangePointsScreen` UI component.
- The submodule `sjy-build-logic` has been updated to reflect these changes.

+++ core/src/main/java/com/carissa/revibes/core/data/user/model/UserData.kt
@@ -12,7 +12,7 @@ data class UserData(
     val email: String,
     val phoneNumber: String,
     val profilePictureUrl: String,
-    val coins: Int,
+    val coins: Long,
     val role: String,
     val createdAt: String,
     val lastClaimedDate: String?,
